Use smaller datatypes
Change uint64_t to uint32_t. Turns out that this implementation is pretty slow,
and there's no way that we're going to be able to run it quickly on large
datasets. And if we aren't going to test on large datasets, we don't have to
worry about integer overflow.
Using smaller datatypes makes communication a little more efficient.

+++ src/main.cpp
@@ -60,9 +60,9 @@ int main(int argc, char* argv[]) {
 
   if(!myid) fprintf(stdout,"Begin suffix array construction\n");
 
-  uint64_t* suffixarray = NULL;
+  uint32_t* suffixarray = NULL;
   try {
-    suffixarray = new uint64_t[size]();
+    suffixarray = new uint32_t[size]();
   } catch (std::bad_alloc& ba) {
     fprintf(stderr, "Bad alloc \n");
     MPI_Finalize();
